move to user service and catch strange casees
