+++ projects/core/src/lib/util/components/backend-status-page/backend-status-page.component.scss
@@ -10,3 +10,9 @@
 .error {
     color: var(--geoengine-warn-color);
 }
+
+mat-card-header {
+    ::ng-deep .mat-card-header-text {
+        width: 100%;
+    }
+}
